change types to float

+++ lib/point.c
@@ -32,8 +32,8 @@ void setCityCoordsAndName(City * c, float x, float y, char name[50]) {
 }
 
 float getDistanceBetween(City * c1, City * c2) {
-    double dx = c2->x - c1->x;
-    double dy = c2->y - c1->y;
-    return (float)sqrt(dx*dx + dy*dy);
+    float dx = c2->x - c1->x;
+    float dy = c2->y - c1->y;
+    return sqrt(dx*dx + dy*dy);
 }
